+++ Source/Lib/Codec/EbComputeSAD.h
@@ -91,14 +91,14 @@ extern "C" {
         uint32_t   mv);
 
     typedef uint32_t(*CombinedAveragingSsd)(
-        uint8_t  *src,
-        uint32_t  src_stride,
-        uint8_t  *ref1,
-        uint32_t  ref1_stride,
-        uint8_t  *ref2,
-        uint32_t  ref2_stride,
-        uint32_t  height,
-        uint32_t  width
+        uint8_t   *src,
+        ptrdiff_t  src_stride,
+        uint8_t   *ref1,
+        ptrdiff_t  ref1_stride,
+        uint8_t   *ref2,
+        ptrdiff_t  ref2_stride,
+        uint32_t   height,
+        uint32_t   width
         );
 
     /***************************************
@@ -223,14 +223,14 @@ extern "C" {
     };
 
     uint32_t combined_averaging_ssd_c(
-        uint8_t  *src,
-        uint32_t  src_stride,
-        uint8_t  *ref1,
-        uint32_t  ref1_stride,
-        uint8_t  *ref2,
-        uint32_t  ref2_stride,
-        uint32_t  height,
-        uint32_t  width);
+        uint8_t   *src,
+        ptrdiff_t  src_stride,
+        uint8_t   *ref1,
+        ptrdiff_t  ref1_stride,
+        uint8_t   *ref2,
+        ptrdiff_t  ref2_stride,
+        uint32_t   height,
+        uint32_t   width);
 
     static CombinedAveragingSsd FUNC_TABLE combined_averaging_ssd_func_ptr_array[ASM_TYPE_TOTAL] =
     {
